+++ ejercicio_individual/src/dama.rs
@@ -1,12 +1,12 @@
 use crate::pieza::{Pieza, Color};
 
 pub(crate) struct Dama {
-    posicion: (usize,usize),
+    posicion: (i8,i8),
     color: Color
 }
 
 impl Dama{
-    pub(crate) fn new(posicion: (usize, usize), color: Color) -> Self{
+    pub(crate) fn new(posicion: (i8,i8), color: Color) -> Self{
         Dama { color, posicion }
     }
 }
@@ -25,7 +25,7 @@ impl Pieza for Dama {
         return &self.color;
     }
 
-    fn posicion(&self) -> &(usize,usize){
+    fn posicion(&self) -> &(i8,i8){
         return &self.posicion;
     }
 }
