+++ ejercicio_individual/src/pieza.rs
@@ -39,10 +39,10 @@ impl PartialEq for Color {
 pub trait Pieza {
     fn puede_capturar(&self, other: &dyn Pieza) -> bool;
     fn color(&self) -> &Color;
-    fn posicion(&self) -> &(usize,usize);
+    fn posicion(&self) -> &(i8,i8);
 }
 
-pub fn crear_pieza(posicion: (usize,usize), pieza: char) -> Result< Box<dyn Pieza>,Error> {
+pub fn crear_pieza(posicion: (i8,i8), pieza: char) -> Result< Box<dyn Pieza>,Error> {
     let color : Color = if pieza.is_uppercase() { Color::Blanco} else { Color::Negro } ;
 
     match pieza {
